Create Create 10 - proposing a lower Quanta emission
As came up in the qrl discord: I like the idea of a lower quanta emisson.
Like every cryptocurrency, in the beginning there are much more coins mined/distributed than necessary. To me it seems though as if the emission could easily be a third or half of what it currently is.

So I´m proposing to mine 2 Quanta every minute instead of close to 6 right now and  keep the current exponential reduction.